+++ packages/calendar/src/js/index.ts
@@ -1,0 +1,3 @@
+export * from './Calendar';
+export * from './mockData';
+export * from '../trads/i18n';